Fixes to the compliation environment, motivated by the need to compile on MacOS X
git-svn-id: https://scm.ti-edu.ch/reposvn/opensmt2/opensmt-dev/trunk@562 d17752b6-1f7d-11e2-a792-abcf70c528dd

+++ src/logics/LRALogic.C
@@ -954,7 +954,7 @@ void LRALogic::deserializeLogicData(const int* logicdata_buf)
     assert(mydata_init < logicdata_buf[0]); // Inside the buffer still
     int sz = logicdata_buf[mydata_init];
     for (int i = 0; i < sz; i++) {
-        SymRef sr = {logicdata_buf[mydata_init+1+i]};
+        SymRef sr = {(uint32_t) logicdata_buf[mydata_init+1+i]};
         SymId id = sym_store[sr].getId();
         for (int j = reals.size(); j <= id; j++)
             reals.push(NULL);
